Improved: JTDateEpochMappings, the didMapObjectFromJSON callback, and made using inheritence with mappings easier.
Now we send the original JTValidJSONResponse (dict) object with the
didMapObjectFromJSON callback. (This is convenient for basic operations
or properties that need to be altered before the mapping is supposed to
be used.
Changed JTMappings.mapping from an NSDictionary into an
NSMutableDictionary to make it easier to use with inheritence. (You can
call the super method and just add items to the .mapping mutable
dictionary instead of having to copy and replace it.)
Added better warning message for unhandled mapping.
Fixed bug with JTDateEpochMappings where a rounding error could cause
you to be one second off.
JTDateEpochMappings now works with NSString's parsed from the JSON.
(Previously it worked only with NSNumber.)
Removed unnecessary NSObject* casts. (Did you put those in because of
compiler warnings?)